+++ src/main/kotlin/nl/sugcube/dirtyarrows/bow/ability/ExplodingBow.kt
@@ -1,0 +1,40 @@
+package nl.sugcube.dirtyarrows.bow.ability
+
+import nl.sugcube.dirtyarrows.DirtyArrows
+import nl.sugcube.dirtyarrows.bow.BowAbility
+import nl.sugcube.dirtyarrows.bow.DefaultBow
+import nl.sugcube.dirtyarrows.util.createExplosion
+import nl.sugcube.dirtyarrows.util.showSmokeParticle
+import org.bukkit.Material
+import org.bukkit.entity.Arrow
+import org.bukkit.entity.Player
+import org.bukkit.event.entity.ProjectileHitEvent
+import org.bukkit.event.entity.ProjectileLaunchEvent
+import org.bukkit.inventory.ItemStack
+
+/**
+ * Shoots arrows that explode on impact.
+ *
+ * @author SugarCaney
+ */
+open class ExplodingBow(plugin: DirtyArrows) : BowAbility(
+        plugin = plugin,
+        type = DefaultBow.EXPLODING,
+        canShootInProtectedRegions = false,
+        protectionRange = 5.0,
+        costRequirements = listOf(ItemStack(Material.TNT, 1)),
+) {
+
+    override fun launch(player: Player, arrow: Arrow, event: ProjectileLaunchEvent) {
+        player.consumeBowItems()
+    }
+
+    override fun land(arrow: Arrow, player: Player, event: ProjectileHitEvent) {
+        arrow.location.createExplosion(power = 4.0f)
+        arrow.remove()
+    }
+
+    override fun particle(tickNumber: Int) = arrows.forEach {
+        it.showSmokeParticle()
+    }
+}